ApiRegistry updates for Google Recaptcha and AddressLookup components
