+++ src/main/java/gwt/material/design/incubator/client/google/addresslookup/AddressLookup.java
@@ -19,9 +19,6 @@
  */
 package gwt.material.design.incubator.client.google.addresslookup;
 
-import com.google.gwt.core.client.Callback;
-import com.google.gwt.core.client.GWT;
-import com.google.gwt.core.client.ScriptInjector;
 import com.google.gwt.event.shared.HandlerRegistration;
 import gwt.material.design.client.ui.MaterialTextBox;
 import gwt.material.design.incubator.client.google.addresslookup.constants.AddressComponentType;
@@ -30,24 +27,56 @@ import gwt.material.design.incubator.client.google.addresslookup.events.AddressL
 import gwt.material.design.incubator.client.google.addresslookup.events.HasAddressLookupHandlers;
 import gwt.material.design.incubator.client.google.addresslookup.events.PlaceChangedEvent;
 import gwt.material.design.incubator.client.google.addresslookup.js.JsAddressLookup;
-import gwt.material.design.incubator.client.google.addresslookup.js.options.*;
+import gwt.material.design.incubator.client.google.addresslookup.js.options.AddressLookupOptions;
+import gwt.material.design.incubator.client.google.addresslookup.js.options.ComponentRestrictions;
+import gwt.material.design.incubator.client.google.addresslookup.js.options.LatLngBounds;
+import gwt.material.design.incubator.client.google.addresslookup.js.options.PlaceResult;
 import gwt.material.design.incubator.client.google.addresslookup.js.options.result.GeocoderAddressComponent;
-import gwt.material.design.jscore.client.api.Navigator;
 
 /**
  * A service to provide Place predictions based on a user's text input. It attaches to an input element of type text,
  * and listens for text entry in that field. The list of predictions is presented as a drop-down list, and is updated
  * as text is entered.
+ * <p>
+ * <h3>UiBinder</h3>
+ * <pre>
+ * {@code
+ *      <google:addresslookup.AddressLookup grid="s12"  ui:field="addressLookup" fieldType="OUTLINED" label="Address" placeholder="Enter any location..."/>
+ * }
+ * </pre>
+ * <h3>Java - Loading the API</h3>
+ * <pre>
+ * {@code
+ *  @literal @UiField
+ *      AddressLookup addressLookup;
+ *
+ *      // Loading API
+ *      ApiRegistry.register(new AddressLookupApi("AIzaSyCcFsjlqr-DR6acrZ8xZKhXNGxeS3nDmIE"), new Callback<Void, Exception>() {
+ *         @literal @Override
+ *          public void onFailure(Exception exception) {
+ *              MaterialToast.fireToast(exception.getMessage());
+ *          }
+ *
+ *         @literal @Override
+ *          public void onSuccess(Void aVoid) {
+ *              addressLookup.load();
+ *          }
+ *      });
+ * }
+ * </pre>
+ * <p>
+ * <p><i>
+ * Note: This component is under the incubation process and subject to change.
+ * </i></p>
  *
  * @author kevzlou7979
  * @see <a href="https://developers.google.com/maps/documentation/javascript/reference/3/#Autocomplete">AddressLookup</a>
  */
-public class AddressLookup extends MaterialTextBox implements HasAddressLookupHandlers {
+public class AddressLookup extends MaterialTextBox
+        implements HasAddressLookupHandlers {
 
-    private String apiKey;
     private JsAddressLookup addressLookup;
     private AddressLookupOptions options;
-    private boolean isApiLoaded;
 
     public AddressLookup() {
         setPlaceholder("");
@@ -62,45 +91,9 @@ public class AddressLookup extends MaterialTextBox implements HasAddressLookupHa
         setOptions(options);
     }
 
-    @Override
-    protected void onLoad() {
-        super.onLoad();
-
-        load();
-    }
-
-    protected void load() {
-        if (!isApiLoaded) {
-            ScriptInjector.fromUrl("https://maps.googleapis.com/maps/api/js?key=" + apiKey + "&libraries=places")
-                    .setWindow(ScriptInjector.TOP_WINDOW)
-                    .setCallback(new Callback<Void, Exception>() {
-                        @Override
-                        public void onFailure(Exception e) {
-                            GWT.log(e.getMessage(), e);
-                        }
-
-                        @Override
-                        public void onSuccess(Void aVoid) {
-                            addressLookup = new JsAddressLookup(valueBoxBase.getElement(), options);
-                            addressLookup.addListener(AddressLookupEvents.PLACE_CHANGED, () -> PlaceChangedEvent.fire(AddressLookup.this));
-                        }
-                    }).inject();
-            isApiLoaded = true;
-        }
-    }
-
-    /**
-     * Sets the GCP Place Api Key
-     */
-    public void setApiKey(String apiKey) {
-        this.apiKey = apiKey;
-    }
-
-    /**
-     * Gets the GCP Place API Key
-     */
-    public String getApiKey() {
-        return apiKey;
+    public void load() {
+        addressLookup = new JsAddressLookup(valueBoxBase.getElement(), options);
+        addressLookup.addListener(AddressLookupEvents.PLACE_CHANGED, () -> PlaceChangedEvent.fire(AddressLookup.this));
     }
 
     /**
@@ -169,13 +162,6 @@ public class AddressLookup extends MaterialTextBox implements HasAddressLookupHa
             }
         }
         return null;
-    }
-
-    /**
-     * Will check if the Google Place API has been loaded
-     */
-    public boolean isApiLoaded() {
-        return isApiLoaded;
     }
 
     @Override
