+++ src/main/java/gwt/material/design/incubator/client/google/addresslookup/js/options/Circle.java
@@ -2,7 +2,7 @@
  * #%L
  * GwtMaterial
  * %%
- * Copyright (C) 2015 - 2017 GwtMaterialDesign
+ * Copyright (C) 2015 - 2018 GwtMaterialDesign
  * %%
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,20 +17,16 @@
  * limitations under the License.
  * #L%
  */
-package gwt.material.design.incubator.client.recaptcha;
+package gwt.material.design.incubator.client.google.addresslookup.js.options;
 
-import com.google.gwt.core.client.GWT;
-import com.google.gwt.resources.client.ClientBundle;
-import com.google.gwt.resources.client.TextResource;
+import jsinterop.annotations.*;
 
-/**
- * Client Bundle resources for ReCaptcha component<br>
- *
- * @author kevzlou7979
- */
-public interface ReCaptchaDebugClientBundle extends ClientBundle {
-    ReCaptchaDebugClientBundle INSTANCE = GWT.create(ReCaptchaDebugClientBundle.class);
+@JsType(isNative = true, namespace = "google.maps", name = "Circle")
+public class Circle {
 
-    @Source("resources/js/recaptcha.min.js")
-    TextResource recaptchaDebugJs();
+    public Circle(GeoLatLng center, double radius) {
+    }
+
+    @JsMethod
+    public native LatLngBounds getBounds();
 }
