--- src/main/java/gwt/material/design/incubator/client/recaptcha/js/JsReCaptcha.java
@@ -1,41 +1,0 @@
-/*
- * #%L
- * GwtMaterial
- * %%
- * Copyright (C) 2015 - 2017 GwtMaterialDesign
- * %%
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * 
- *      http://www.apache.org/licenses/LICENSE-2.0
- * 
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- * #L%
- */
-package gwt.material.design.incubator.client.recaptcha.js;
-
-import gwt.material.design.jquery.client.api.Functions;
-import gwt.material.design.jquery.client.api.JQueryElement;
-import jsinterop.annotations.JsMethod;
-import jsinterop.annotations.JsPackage;
-import jsinterop.annotations.JsType;
-
-/**
- * JSInterop utils for ReCaptcha component
- *
- * @author kevzlou7979
- */
-@JsType(isNative = true, namespace = JsPackage.GLOBAL)
-public class JsReCaptcha extends JQueryElement {
-
-    @JsMethod(namespace = JsPackage.GLOBAL)
-    public static native JsReCaptcha initReCaptcha(String element, String siteKey, Functions.Func callback, String theme, String type);
-
-    @JsMethod(namespace = JsPackage.GLOBAL)
-    public static native void reset();
-}
