+++ src/main/resources/gwt/material/design/incubator/client/google/recaptcha/resources/js/recaptcha.js
