+++ src/test/java/gwt/material/design/incubator/client/ui/ReCaptchaTest.java
@@ -21,9 +21,9 @@ package gwt.material.design.incubator.client.ui;
 
 import com.google.gwt.dom.client.ScriptElement;
 import gwt.material.design.incubator.client.MaterialWidgetTest;
-import gwt.material.design.incubator.client.recaptcha.ReCaptcha;
-import gwt.material.design.incubator.client.recaptcha.constants.CaptchaTheme;
-import gwt.material.design.incubator.client.recaptcha.constants.CaptchaType;
+import gwt.material.design.incubator.client.google.recaptcha.ReCaptcha;
+import gwt.material.design.incubator.client.google.recaptcha.constants.RecaptchaTheme;
+import gwt.material.design.incubator.client.google.recaptcha.constants.RecaptchaType;
 import gwt.material.design.jquery.client.api.Functions;
 
 /**
@@ -53,10 +53,10 @@ public class ReCaptchaTest extends MaterialWidgetTest<ReCaptcha> {
         ReCaptcha reCaptcha = getWidget();
 
         // Check Themes
-        reCaptcha.setTheme(CaptchaTheme.LIGHT);
-        assertEquals(reCaptcha.getTheme(), CaptchaTheme.LIGHT);
-        reCaptcha.setTheme(CaptchaTheme.DARK);
-        assertEquals(reCaptcha.getTheme(), CaptchaTheme.DARK);
+        reCaptcha.setTheme(RecaptchaTheme.LIGHT);
+        assertEquals(reCaptcha.getTheme(), RecaptchaTheme.LIGHT);
+        reCaptcha.setTheme(RecaptchaTheme.DARK);
+        assertEquals(reCaptcha.getTheme(), RecaptchaTheme.DARK);
 
         // Check Site Key
         final String SITE_KEY = "sampleSitekKey";
@@ -70,10 +70,10 @@ public class ReCaptchaTest extends MaterialWidgetTest<ReCaptcha> {
 
         // Check Type
         // Expected by default - IMAGE
-        assertEquals(reCaptcha.getType(), CaptchaType.IMAGE);
-        reCaptcha.setType(CaptchaType.AUDIO);
-        assertEquals(reCaptcha.getType(), CaptchaType.AUDIO);
-        reCaptcha.setType(CaptchaType.IMAGE);
-        assertEquals(reCaptcha.getType(), CaptchaType.IMAGE);
+        assertEquals(reCaptcha.getType(), RecaptchaType.IMAGE);
+        reCaptcha.setType(RecaptchaType.AUDIO);
+        assertEquals(reCaptcha.getType(), RecaptchaType.AUDIO);
+        reCaptcha.setType(RecaptchaType.IMAGE);
+        assertEquals(reCaptcha.getType(), RecaptchaType.IMAGE);
     }
 }
